+++ R/modelsum.R
@@ -189,13 +189,12 @@ modelsum <- function(formula,  family="gaussian", data, adjust=NULL, na.action =
 
           results <- modelsum_guts(family.list, temp.call, envir = parent.frame(), conf.level = control$conf.level)
 
-          currCols <- maindf[attr(Terms.x, "factors")[, eff] > 0]
-          nmiss <- sum(rowSums(is.na(currCols))[strata.col == strat] > 0)
+          nmiss <- length(results$fit$na.action)
           xList[[eff]][[adj.i]] <- list(
             coeff=results$coeffTidy,
             glance = c(
               results$modelGlance,
-              N = nrow(maindf) - nmiss,
+              N = sum(strata.col == strat) - nmiss,
               Nmiss = nmiss,
               Nmiss2 = nmiss,
               endpoint=yTerm,